+++ app-config-config/src/index.ts
@@ -19,7 +19,7 @@ import {
 } from '@app-config/node';
 import { markAllValuesAsSecret } from '@app-config/extensions';
 import { defaultExtensions, defaultEnvExtensions } from '@app-config/default-extensions';
-import { loadSchema, JSONSchema, SchemaLoadingOptions } from '@app-config/schema';
+import { loadSchema, JSONSchema, SchemaLoadingOptions, Schema } from '@app-config/schema';
 import { loadMetaConfig, loadExtraParsingExtensions } from '@app-config/meta';
 
 export interface ConfigLoadingOptions {
@@ -49,7 +49,7 @@ export interface LoadedConfiguration {
   /** if loadValidatedConfig, this is the normalized JSON schema that was used for validation */
   schema?: JSONSchema;
   /** if loadValidatedConfig, this is the raw AJV validation function */
-  validationFunctionCode?(): string;
+  validationFunctionCode?: Schema['validationFunctionCode'];
 }
 
 export async function loadUnvalidatedConfig({
@@ -195,23 +195,21 @@ export async function loadValidatedConfig(
   options?: ConfigLoadingOptions,
   schemaOptions?: SchemaLoadingOptions,
 ): Promise<LoadedConfiguration> {
-  const [
-    { validate, validationFunctionCode, schema },
-    { fullConfig, parsed, ...rest },
-  ] = await Promise.all([
-    loadSchema({
-      directory: options?.directory,
-      fileNameBase: options?.fileNameBase ? `${options.fileNameBase}.schema` : undefined,
-      environmentVariableName: options?.environmentVariableName
-        ? `${options.environmentVariableName}_SCHEMA`
-        : undefined,
-      environmentOverride: options?.environmentOverride,
-      environmentAliases: options?.environmentAliases,
-      environmentSourceNames: options?.environmentSourceNames,
-      ...schemaOptions,
-    }),
-    loadUnvalidatedConfig(options),
-  ]);
+  const [{ validate, validationFunctionCode, schema }, { fullConfig, parsed, ...rest }] =
+    await Promise.all([
+      loadSchema({
+        directory: options?.directory,
+        fileNameBase: options?.fileNameBase ? `${options.fileNameBase}.schema` : undefined,
+        environmentVariableName: options?.environmentVariableName
+          ? `${options.environmentVariableName}_SCHEMA`
+          : undefined,
+        environmentOverride: options?.environmentOverride,
+        environmentAliases: options?.environmentAliases,
+        environmentSourceNames: options?.environmentSourceNames,
+        ...schemaOptions,
+      }),
+      loadUnvalidatedConfig(options),
+    ]);
 
   if (!isObject(fullConfig)) {
     throw new WasNotObject('Configuration was not an object');