+++ app-config-encryption/src/encryption.ts
@@ -808,7 +808,7 @@ function addForEnvironment<T>(
 }
 
 function decodeTypedArray(buf: ArrayBuffer): string {
-  return String.fromCharCode.apply(null, (new Uint16Array(buf) as any) as number[]);
+  return String.fromCharCode.apply(null, new Uint16Array(buf) as any as number[]);
 }
 
 function stringAsTypedArray(str: string): Uint16Array {