+++ examples/rollup-project/src/index.js
@@ -1,0 +1,3 @@
+import { config } from '@app-config/main';
+
+console.log(config);